README: Cull irrelevant bugzilla-operator docs
I'd borrowed the README content and license from the bugzilla-operator
repository's:

* mfojtik/bugzilla-operator@509dddb656 (add README.md, 2020-05-14).
* mfojtik/bugzilla-operator@464ac890cd (get license, 2020-05-14)

because a bunch of code in this repository is based on an earlier
fork-ish from the bugzilla-operator repository.  But this repository
does not have the LifecyleStale management that bugzilla-operator
does, and it also has additional tools like upcoming-sprint-stats and
whatnot.  I don't understand this repository well enough to describe
everything it's trying to do, but I do expect that Eric did not intend
to diverge the licensing Michal was using upstream-ish.