Better control of subcomponents.
Allow ignoring teams entirely in overrides.
Allow defining teams as the 'default' of the component. Before if a new subcomponent was added those bugs went to UNKNONW
Pull config both from the main shiftzilla file and from a second file about subcomponents. So we use the local 'override' very very little.
Fix SLOs to be absolute instead of 0 if there are unknown team members