blockerslack: split messages longer than slack can handle
Today slack will split messages longer than about 4k characters. It does so whereever
it wants. And if this is in the middle of a link you get a wall of text instead of
something you can click. With enough bugs teams hit the limit. This splits lines on
our side instead of letting slack do it on their side. If a single line ever becomes
longer than 4k slack will split it, but if it's a multiline message this will do good.