bug-automation: add automation to redeploy on changes in github
This creates a 'git-build-watcher' which watches github for changes and
then kicks off a buildconfig if it sees changes.

The buildconfig puts the result into the bug-automation imagestream.

Since the actual thing running the image is a cronjob we don't need a
trigger on the deployment like most things. The next time the cronjob
runs it'll get the new image.